+++ RELEASE-NOTES.md
@@ -73,10 +73,11 @@ Initial version with:
 ## [0.13.0](https://github.com/kisscool62/io.nagurea.smsupsdk/compare/v0.12.0...v0.13.0) (2022-08-15)
 * Webhooks > Create a webhook
 
-## [0.13.1](https://github.com/kisscool62/io.nagurea.smsupsdk/compare/v0.13.0...v0.13.1) (2022-08-16)
+## [0.13.1](https://github.com/kisscool62/io.nagurea.smsupsdk/compare/v0.13.0...v0.13.1) (2022-08-17)
 * Webhooks > Retrieve a webhook
 * Webhooks > Retrieve a single webhook
 * Webhooks > Delete a webhook
+* Webhooks > Update a webhook
 
 ### Upgrade Steps
 * [NO ACTION REQUIRED]
@@ -88,6 +89,7 @@ Initial version with:
 * Webhooks > Retrieve a webhook
 * Webhooks > Retrieve a single webhook
 * Webhooks > Delete a webhook
+* Webhooks > Update a webhook
 
 ### Bug Fixes